extracting pydantic base models to models.py
