+++ STACpopulator/stac_utils.py
@@ -1,27 +1,12 @@
-import re
 import json
-import datetime as dt
-from enum import Enum, auto
-from typing import Any, Iterator, MutableMapping, Optional, Tuple, Union
-from typing import Any, Dict, List, Literal, MutableMapping
-from typing_extensions import TypedDict
+import re
+from typing import Any, Literal, MutableMapping
+
 import pystac
-from pystac.extensions.datacube import Dimension, DimensionType, VariableType, Variable, DatacubeExtension
-from pydantic import AnyHttpUrl, BaseModel, field_validator, Field, ConfigDict, RootModel, AnyUrl
-from urllib.parse import urljoin
+from pystac.extensions.datacube import Dimension, DimensionType, Variable, VariableType
 
+from STACpopulator.models import STACItem, STACItemProperties
 
-import pyessv
-
-
-try:
-    from enum import EnumType as enumtype
-except ImportError:
-    # < Python 3.11
-    from enum import EnumMeta as enumtype
-
-
-STAC_VERSION = "1.0.0"
 
 def url_validate(target: str) -> bool:
     """Validate whether a supplied URL is reliably written.
@@ -48,89 +33,8 @@ def url_validate(target: str) -> bool:
 
 
 def collection2literal(collection):
-    import typing
     terms = tuple(term.label for term in collection)
-    return typing.Literal[terms]
-
-
-class AutoValueEnum(Enum):
-    def _generate_next_value_(  # type: ignore
-        name: str, start: int, count: int, last_values: List[Any]
-    ) -> Any:
-        return name
-
-
-# DH: There is a question here whether we want to use pystac.Item or not.
-# pystac.Item takes datetime, start_datetime and end_datetime as optional parameters, and then copies them into
-# properties.
-# If we use pystac.Item, we don't have to put start_datetime and end_datetime into Properties, we can let pystac do
-# that.
-class ItemProperties(BaseModel):
-    start_datetime: Optional[dt.datetime] = None
-    end_datetime: Optional[dt.datetime] = None
-    datetime: Optional[dt.datetime] = None
-
-    @field_validator("datetime", mode="before")
-    def validate_datetime(cls, v: Union[dt.datetime, str], values: Dict[str, Any]) -> dt:
-        if v == "null":
-            if not values["start_datetime"] and not values["end_datetime"]:
-                raise ValueError(
-                    "start_datetime and end_datetime must be specified when datetime is null"
-                )
-
-
-class Geometry(TypedDict):
-    type: str
-    coordinates: List[List[List[float]]]
-
-class Asset(BaseModel):
-    href: AnyHttpUrl
-    media_type: Optional[str] = None
-    title: Optional[str] = None
-    description: Optional[str] = None
-    roles: Optional[List[str]] = None
-
-class Link(BaseModel):
-    """
-    https://github.com/radiantearth/stac-spec/blob/v1.0.0/collection-spec/collection-spec.md#link-object
-    """
-
-    href: str = Field(..., alias="href", min_length=1)
-    rel: str = Field(..., alias="rel", min_length=1)
-    type: Optional[str] = None
-    title: Optional[str] = None
-    # Label extension
-    label: Optional[str] = Field(None, alias="label:assets")
-    model_config = ConfigDict(use_enum_values=True)
-
-    def resolve(self, base_url: str) -> None:
-        """resolve a link to the given base URL"""
-        self.href = urljoin(base_url, self.href)
-
-
-class PaginationLink(Link):
-    """
-    https://github.com/radiantearth/stac-api-spec/blob/master/api-spec.md#paging-extension
-    """
-
-    rel: Literal["next", "previous"]
-    method: Literal["GET", "POST"]
-    body: Optional[Dict[Any, Any]] = None
-    merge: bool = False
-
-Links = RootModel[List[Union[PaginationLink, Link]]]
-
-
-class Item(BaseModel):
-    id: str = Field(..., alias="id", min_length=1)
-    geometry: Optional[Geometry] = None
-    bbox: Optional[List[float]] = None
-    properties: Optional[ItemProperties] = None
-    assets: Dict[str, Asset] = None
-    stac_extensions: Optional[List[AnyUrl]] = []
-    collection: Optional[str] = None
-    datetime: Optional[dt.datetime] = None  # Not in the spec, but needed by pystac.Item.
-
+    return Literal[terms]
 
 
 class CFJsonItem:
@@ -162,9 +66,9 @@ class CFJsonItem:
         }
 
         # Validate using pydantic data model if given
-        datamodel = datamodel or dict
+        datamodel = datamodel or STACItemProperties
 
-        class MySTACItem(Item):
+        class MySTACItem(STACItem):
             properties: datamodel
 
         # Create STAC item
